fix(clips): harden raw-cut generation mode
