+++ apps/server/api/src/collections/clip-projects/clip-projects.controller.spec.ts
@@ -1,3 +1,4 @@
+import { ClipProjectHandoffsController } from '@api/collections/clip-projects/clip-project-handoffs.controller';
 import { ClipProjectsController } from '@api/collections/clip-projects/clip-projects.controller';
 import type { ClipProjectsService } from '@api/collections/clip-projects/clip-projects.service';
 import { CreateClipProjectFromYoutubeDto } from '@api/collections/clip-projects/dto/create-clip-project-from-youtube.dto';
@@ -87,6 +88,7 @@ describe('ClipProjectsController', () => {
   };
 
   let controller: ClipProjectsController;
+  let handoffsController: ClipProjectHandoffsController;
   let clipProjectsService: ReturnType<typeof createMockClipProjectsService>;
   let clipGenerationService: ReturnType<typeof createMockClipGenerationService>;
   let clipFactoryQueueService: { enqueue: ReturnType<typeof vi.fn> };
@@ -130,6 +132,10 @@ describe('ClipProjectsController', () => {
       clipGenerationService as ClipGenerationService,
       { rewrite: vi.fn() } as unknown as HighlightRewriteService,
       creditsUtilsService as unknown as CreditsUtilsService,
+    );
+    handoffsController = new ClipProjectHandoffsController(
+      createMockLogger(),
+      clipProjectsService as ClipProjectsService,
       clipResultsService as unknown as ClipResultsService,
       editorProjectsService as unknown as EditorProjectsService,
       publishHandoffService as unknown as PublishHandoffService,
@@ -164,6 +170,7 @@ describe('ClipProjectsController', () => {
           organization: expect.any(String),
           settings: expect.objectContaining({
             maxClips: 12,
+            mode: 'avatar',
           }),
           sourceVideoUrl: dto.youtubeUrl,
           user: expect.any(String),
@@ -226,6 +233,11 @@ describe('ClipProjectsController', () => {
           voiceId: undefined,
         }),
       );
+      expect(clipProjectsService.create).toHaveBeenCalledWith(
+        expect.objectContaining({
+          settings: expect.objectContaining({ mode: 'raw-cut' }),
+        }),
+      );
     });
   });
 
@@ -249,6 +261,30 @@ describe('ClipProjectsController', () => {
       });
 
       expect(validateSync(dto)).toEqual([]);
+    });
+
+    it('should validate optional avatar credentials in raw-cut mode', () => {
+      const dto = plainToInstance(CreateClipProjectFromYoutubeDto, {
+        avatarId: 123,
+        mode: 'raw-cut',
+        voiceId: false,
+        youtubeUrl: 'https://youtu.be/dQw4w9WgXcQ',
+      });
+
+      const errors = validateSync(dto);
+
+      expect(errors.map((error) => error.property)).toEqual(
+        expect.arrayContaining(['avatarId', 'voiceId']),
+      );
+    });
+
+    it('should reject unknown generation modes', () => {
+      const dto = plainToInstance(CreateClipProjectFromYoutubeDto, {
+        mode: 'unknown',
+        youtubeUrl: 'https://youtu.be/dQw4w9WgXcQ',
+      });
+
+      expect(validateSync(dto).map((error) => error.property)).toContain('mode');
     });
 
     it('should reject non-youtube URLs', () => {
@@ -333,6 +369,32 @@ describe('ClipProjectsController', () => {
       expect(validateSync(dto)).toEqual([]);
     });
 
+    it('should validate optional avatar credentials in raw-cut mode', () => {
+      const dto = plainToInstance(GenerateClipsDto, {
+        avatarId: 123,
+        editedHighlights,
+        mode: 'raw-cut',
+        selectedHighlightIds: ['highlight-1'],
+        voiceId: false,
+      });
+
+      const errors = validateSync(dto);
+
+      expect(errors.map((error) => error.property)).toEqual(
+        expect.arrayContaining(['avatarId', 'voiceId']),
+      );
+    });
+
+    it('should reject unknown generation modes', () => {
+      const dto = plainToInstance(GenerateClipsDto, {
+        editedHighlights,
+        mode: 'unknown',
+        selectedHighlightIds: ['highlight-1'],
+      });
+
+      expect(validateSync(dto).map((error) => error.property)).toContain('mode');
+    });
+
     it.each([
       'did',
       'tavus',
@@ -399,6 +461,7 @@ describe('ClipProjectsController', () => {
           }),
         ],
         progress: 0,
+        settings: expect.objectContaining({ mode: 'avatar' }),
         status: 'generating',
       }),
     );
@@ -509,6 +572,30 @@ describe('ClipProjectsController', () => {
     );
   });
 
+  it('should reject raw-cut generation when the project has no source video', async () => {
+    const project = createProject(projectId, organizationId);
+    const dto: GenerateClipsDto = {
+      editedHighlights: [
+        {
+          id: 'highlight-1',
+          summary: 'Edited summary',
+          title: 'Edited title',
+        },
+      ],
+      mode: 'raw-cut',
+      selectedHighlightIds: ['highlight-1'],
+    };
+
+    vi.mocked(clipProjectsService.findOne).mockResolvedValue(project);
+
+    await expect(
+      controller.generateClips(currentUser as never, projectId, dto),
+    ).rejects.toThrow('requires a source video');
+
+    expect(clipProjectsService.patch).not.toHaveBeenCalled();
+    expect(clipGenerationService.generateClips).not.toHaveBeenCalled();
+  });
+
   it('creates an editor handoff for a ready clip result', async () => {
     const project = createProject(projectId, organizationId);
     vi.mocked(clipProjectsService.findOne).mockResolvedValue(project);
@@ -532,7 +619,7 @@ describe('ClipProjectsController', () => {
       id: 'editor-project-1',
     });
 
-    const result = await controller.createEditorHandoff(
+    const result = await handoffsController.createEditorHandoff(
       currentUser as never,
       projectId,
       'clip-result-1',
@@ -602,7 +689,7 @@ describe('ClipProjectsController', () => {
       schedule: 'immediate',
     });
 
-    const result = await controller.createPublishHandoff(
+    const result = await handoffsController.createPublishHandoff(
       currentUser as never,
       projectId,
       'clip-result-1',
@@ -646,7 +733,7 @@ describe('ClipProjectsController', () => {
     });
 
     await expect(
-      controller.createPublishHandoff(
+      handoffsController.createPublishHandoff(
         currentUser as never,
         projectId,
         'clip-result-1',
