+++ apps/server/api/src/collections/clip-projects/clip-projects.module.ts
@@ -1,3 +1,4 @@
+import { ClipProjectHandoffsController } from '@api/collections/clip-projects/clip-project-handoffs.controller';
 import { ClipProjectsController } from '@api/collections/clip-projects/clip-projects.controller';
 import { ClipProjectsCoreModule } from '@api/collections/clip-projects/clip-projects-core.module';
 import { ClipResultsModule } from '@api/collections/clip-results/clip-results.module';
@@ -9,7 +10,7 @@ import { ClipOrchestratorModule } from '@api/services/clip-orchestrator/clip-orc
 import { forwardRef, Module } from '@nestjs/common';
 
 @Module({
-  controllers: [ClipProjectsController],
+  controllers: [ClipProjectsController, ClipProjectHandoffsController],
   exports: [ClipProjectsCoreModule],
   imports: [
     forwardRef(() => ClipProjectsCoreModule),
