+++ apps/server/api/src/collections/clip-projects/dto/create-clip-project-from-youtube.dto.ts
@@ -51,7 +51,8 @@ export class CreateClipProjectFromYoutubeDto {
 
   @ValidateIf(
     (dto: CreateClipProjectFromYoutubeDto) =>
-      (dto.mode ?? DEFAULT_CLIP_RESULT_MODE) === 'avatar',
+      (dto.mode ?? DEFAULT_CLIP_RESULT_MODE) === 'avatar' ||
+      dto.avatarId !== undefined,
   )
   @IsString()
   @ApiProperty({
@@ -62,7 +63,8 @@ export class CreateClipProjectFromYoutubeDto {
 
   @ValidateIf(
     (dto: CreateClipProjectFromYoutubeDto) =>
-      (dto.mode ?? DEFAULT_CLIP_RESULT_MODE) === 'avatar',
+      (dto.mode ?? DEFAULT_CLIP_RESULT_MODE) === 'avatar' ||
+      dto.voiceId !== undefined,
   )
   @IsString()
   @ApiProperty({
