+++ packages/queue-contracts/package.json
@@ -1,6 +1,7 @@
 {
   "dependencies": {
-    "@genfeedai/enums": "workspace:*"
+    "@genfeedai/enums": "workspace:*",
+    "@genfeedai/interfaces": "workspace:*"
   },
   "devDependencies": {
     "typescript": "6.0.2"
